Added AAC decoding support. Fixed the source of the major sound-skipping glitch when tuning in and out of stations (it was mistakenly calculating the weight of the pink noise to 0, even though it was still decoding audio from the tuned track. Probably doesn't build non-OS X at the moment.

+++ src/Decoder.h
@@ -32,7 +32,7 @@ public:
 
 	virtual	int			encoding()							const								=	0;
 	virtual	int			numChannels()						const								=	0;
-	virtual	long		rate()								const								=	0;
+	virtual	uint32_t	rate()								const								=	0;
 	
 	virtual	off_t		currentFrame()						const								=	0;
 	virtual	void		setCurrentFrame(off_t inFrame)											=	0;
